# KafkaSinkCluster: fix consumer latency issues