This PR is an extension of the rewrite `fetch.max.wait.ms`/`fetch.min.bytes` solution to the consumer latency issue.

In addition to rewriting those fields on the fetch requests to 1, this PR also implements shotover side handling of the original values of those fields as sent by the client.

So for example client sends fetch request with:
* `fetch.max.wait.ms=500`
* `fetch.min.bytes=100`

Shotover will rewrite both of those fields to 1 before sending the split requests to the kafka brokers.
However shotover will store the 500 and 100 values and abide by them in its communication with the client.

The logic shotover uses to achieve this looks like:
1. send all split requests to their destination with configs rewritten to 1
2. wait for all responses to come back
    * assuming a healthy cluster this should only be a tiny bit slower than waiting for 1 request from 1 broker
3. If the original value of `fetch.max.wait.ms` has expired or the count of all of the bytes of all the records in the responses > `fetch.min.bytes` return the values to the client.
4. Otherwise discard the responses and repeat from step 1 with an exponential backoff sleep, starting at 1ms -> 2ms -> 4ms -> 8ms -> 16ms -> ...

## performance consequences compared to just rewriting the values

* reduce client<->shotover network overhead + encode/decode cpu cost
* when the user sets both `fetch.max.wait.ms` and `fetch.min.bytes` to a high value
   + increase shotover<->broker network overhead + encode/decode cpu cost
* when the user sets both `fetch.max.wait.ms` to a high value and `fetch.min.bytes` to 1.
   + no extra cost
   
This means that with the default configuration of the java and rdkafka clients we get only performance improvements over just rewriting the values.

## Possible further extension

Instead of discarding the responses in step 4 we could hold onto the records we obtained and alter the offset in the fetch requests so that we do not refetch requests that shotover already has. Then we combine the original and refetched responses together.
This would:
* reduce the disk, cpu and network load on the kafka cluster
* probably neutral on shotover load (less network access + decode but more shuffling around of records)
* higher shotover memory usage
  + the higher the batching configuration values the higher the memory usage
* increase complexity of the implementation but probably not by much

# Local benchmark results for this PR:
![image](https://github.com/user-attachments/assets/516a62f3-a431-420a-a42c-6a6d404a313d)